ref(upload): Rename command to upload-dbg
